+++ GServer/Unit-Tests/MessageTest.cs
@@ -29,7 +29,7 @@ namespace Unit_Tests
             ds.Push("hello world");
             ds.Push(13.221F);
             ds.Push(14.32D);
-            var msg = new Message(MessageType.Ack, Mode.Reliable | Mode.Sequenced, ds);
+            var msg = new Message((short)MessageType.Ack, Mode.Reliable | Mode.Sequenced, ds);
             msg.MessageId = 123;
             msg.ConnectionToken = token;
             var bytes = msg.Serialize();
